refactor(sandboxstore): split store responsibilities
